Added multiclick capability? Hopefully?
Further revamped the c++ but, can garentee it still needs work.

And sooner or later i will need to make new functions but, i dont think it needs to be divided up yet...
However i do know, it should be divided between config and usage. So, will do that later.